move SignedTransaction to model module
